+++ Procfile
@@ -1,2 +1,3 @@
 release: python manage.py migrate
 web: gunicorn social.wsgi
+worker: python manage.py process_tasks
